+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Upgraded dependencies
 
 ### Fixed:
+- Entry DateTime modal layout
 
 ## [0.8.276] - 2023-02-28
 ### Changed: